apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.34
psycopg[binary]>=3.2.2
alembic==1.13.2

pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1
passlib==1.7.4
redis==5.0.1

